refactor: split buildDOM, getBuildDOMLibrary into modules
