+++ public/index.html
@@ -112,7 +112,6 @@
   <script src="js/utils/get-username.js"></script>
   <script src="js/services/is-logged-in.js"></script>
   <script type="module" src="js/app.js"></script>
-  <script defer src="js/table/build-dom.js"></script>
   <script src="js/table/create-table-container.js"></script>
   <script src="js/table/create-table-template.js"></script>
   <script src="js/table/create-table.js"></script>
